Feat: Extract areOrderInfoEqual function to a utility file
